feat (*): Refactor the design to use the same primitives everywhere
This commit re-architect the metrics collection such that it uses the
same primitive (gauge, counter) in all points whether the metrics are
collected, rather than needing an additional singleton to handle the
state. This reduces the amount of code both needed to implement a
metric, and for each primitive metric type.

Singletons are still required to keep state for each given metric, so a
new factory method was introduced lazily create and store the metric
types in the Magento registry.

Future work may refactor the factory such that it is a singleton itself,
and thus does not need to use the Magento registry for storage.